requirements.txt
tensorflow==2.8.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.24.0
geopy==2.2.0
Pillow==9.0.1
numpy==1.21.6
pandas==1.3.5

